tutorial/requirements.txt
APScheduler==3.6.3
Flask==1.1.1
requests==2.22.0
SQLAlchemy==1.3.20
seaborn==0.10.1
mpl_finance==0.10.1
pandas_datareader
matplotlib==3.2.1
sshtunnel==0.2.1
selenium==3.141.0
PyMySQL==0.9.3
Faker==2.0.4
interval==1.0.0
fake_useragent==0.1.11
urllib3==1.25.7
redis==3.3.11
beautifulsoup4==4.9.3
cx_Oracle==8.1.0
Pillow==8.1.0
pyzbar==0.1.8
twstock==1.3.1
Django==3.1.7
django-tables2==2.4.0
djangorestframework==3.12.4
Pygments==2.7.2
HTMLTestRunner-Python3==0.8.0
mysqlclient==2.0.2